support hyper, split mods
